next-esim-app/functions/requirements.txt
firebase_functions>=0.1.0
firebase-admin>=6.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
Pillow>=10.4.0
stripe>=7.0.0